"Extend mempool expiry and add locktime handling in transactions"
This commit extends the expiration time for transactions in the memory pool from 26 hours to 168 hours in an effort to provide a more user-friendly process for managing transactions that may not be processed immediately due to network traffic.

We have also added locktime handling for transactions in PocketTransactionRpc.cpp. The locktime allows further control over when a transaction can be added to the blockchain, improving the capabilities of transaction scheduling.

In the validation.cpp file, we have extended the list of content types and adjusted the conditions for transaction acceptance into the memory pool. This change should increase the versatility of our transaction capabilities.

Furthermore, in the txmempool.cpp file, we have adjusted the conditions for transaction removal from the pool, considering the locktime value. This will prevent premature transaction expulsion and ensure a more predictable transaction life cycle.